+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/IosButtons.cs
@@ -1,0 +1,42 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+#if PLATFORM_IOS
+using System.Runtime.InteropServices;
+#endif
+
+public class IosButtons : MonoBehaviour
+{
+    public void ThrowObjectiveC()
+    {
+#if PLATFORM_IOS
+        throwObjectiveC();
+#else
+        Debug.Log("Requires IL2CPP. Try this on a native player that supports Objective-C native plugins.");
+#endif
+    }
+
+#if PLATFORM_IOS
+    // ObjectiveCPlugin.m
+    [DllImport("__Internal")]
+    private static extern void throwObjectiveC();
+#endif
+
+    public void RunOutOfMemory()
+    {
+        Debug.Log("Attempting to run out of memory");
+
+        StartCoroutine(ConsumeMemory());
+    }
+
+    IEnumerator ConsumeMemory()
+    {
+        var memoryEaters = new List<object[]>();
+        while(true)
+        {
+            yield return new WaitForEndOfFrame();
+            memoryEaters.Add(new object[1000000]);
+        }
+    }
+
+}